+++ tests/all.ts
@@ -57,6 +57,7 @@ import { testit as issue5868 } from "./issue-5868.ts";
 import { testit as issue5870 } from "./issue-5870.ts";
 import { testit as issue5870ListDirs } from "./issue-5870-list-dirs.ts";
 import { testit as issue5869 } from "./issue-5869.ts";
+import { testit as issue5873 } from "./issue-5873.ts";
 import { testit as reloadDebounce } from "./reload-debounce.ts";
 import { testit as parseTipBrackets } from "./parse-tip-brackets.ts";
 import { testit as ghsaP2ph2rvmQ37m } from "./security-ghsa-p2ph-2rvm-q37m.ts";
@@ -110,6 +111,7 @@ const tests: [string, () => void | Promise<void>][] = [
   ["issue-5870", issue5870],
   ["issue-5870-list-dirs", issue5870ListDirs],
   ["issue-5869", issue5869],
+  ["issue-5873", issue5873],
   ["reload-debounce", reloadDebounce],
   ["parse-tip-brackets", parseTipBrackets],
   ["security-ghsa-p2ph-2rvm-q37m", ghsaP2ph2rvmQ37m],